requirements.txt
# ResolvaBot LLM - Complete Dependencies

# Core
numpy>=1.24.0
pandas>=2.0.0

# PDF Extraction
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Text Processing & NLP
nltk>=3.8.1

# Embeddings & ML
sentence-transformers>=2.2.2
torch>=2.0.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4

# BM25 / Keyword Index
whoosh>=2.7.4

# OpenAI LLM (modern SDK)
openai>=1.0.0

# Wikipedia Fallback
wikipedia-api>=0.6.0

# Environment Variables
python-dotenv>=1.0.0

# Web UI
streamlit>=1.28.0
